+++ cmake/AddBipedalLocomotionUnitTest.cmake
@@ -5,6 +5,29 @@
 # This software may be modified and distributed under the terms of the
 # BSD-3-Clause license. See the accompanying LICENSE file for details.
 
+# This software may be modified and distributed under the terms of the
+# GNU Lesser General Public License v2.1 or any later version.
+
+# Fetch catch2 only if the testing are built
+if (BUILD_TESTING)
+  find_package(Catch2 3.0.1 QUIET)
+  cmake_dependent_option(USE_SYSTEM_Catch2 "Use system Catch2" ON "Catch2_FOUND" OFF)
+  if(NOT USE_SYSTEM_Catch2)
+    include(FetchContent)
+    FetchContent_Declare(Catch2
+      GIT_REPOSITORY https://github.com/catchorg/Catch2.git
+      GIT_TAG        v3.0.1)
+
+    FetchContent_GetProperties(Catch2)
+    if(NOT Catch2_POPULATED)
+      message(STATUS "Fetching Catch2...")
+      FetchContent_MakeAvailable(Catch2)
+    endif()
+  endif()
+endif()
+
+
+
 if (FRAMEWORK_RUN_Valgrind_tests)
     set(CTEST_MEMORYCHECK_COMMAND ${VALGRIND_PROGRAM})
     set(MEMORYCHECK_COMMAND ${VALGRIND_PROGRAM})
@@ -21,16 +44,9 @@ if (FRAMEWORK_RUN_Valgrind_tests)
     separate_arguments(MEMCHECK_COMMAND_COMPLETE)
 endif()
 
-if (FRAMEWORK_COMPILE_tests)
-    configure_file(cmake/Catch2Main.cpp.in ${CMAKE_BINARY_DIR}/Testing/Catch2Main.cpp)
-    add_library(CatchTestMain ${CMAKE_BINARY_DIR}/Testing/Catch2Main.cpp)
-    target_link_libraries(CatchTestMain PUBLIC Catch2::Catch2)
-endif()
-
-
 function(add_bipedal_test)
 
-    if(FRAMEWORK_COMPILE_tests)
+    if(BUILD_TESTING)
 
       set(options )
       set(oneValueArgs NAME)
@@ -51,7 +67,7 @@ function(add_bipedal_test)
       add_executable(${targetname}
           "${unit_test_files}")
 
-      target_link_libraries(${targetname} PRIVATE CatchTestMain ${${prefix}_LINKS})
+      target_link_libraries(${targetname} PRIVATE Catch2::Catch2WithMain ${${prefix}_LINKS})
       target_compile_definitions(${targetname} PRIVATE CATCH_CONFIG_FAST_COMPILE CATCH_CONFIG_DISABLE_MATCHERS CATCH_CONFIG_ENABLE_CHRONO_STRINGMAKER)
       target_compile_features(${targetname} PUBLIC cxx_std_17)
       target_compile_definitions(${targetname} PRIVATE -D_USE_MATH_DEFINES)