+++ src/ContactModels/tests/ContinousContactModelTest.cpp
@@ -8,7 +8,7 @@
 #include <random>
 
 // Catch2
-#include <catch2/catch.hpp>
+#include <catch2/catch_test_macros.hpp>
 
 #include <iDynTree/Core/EigenHelpers.h>
 #include <iDynTree/Core/SpatialAcc.h>