+++ src/Contacts/tests/ContactDetectors/SchmittTriggerDetectorUnitTest.cpp
@@ -7,7 +7,7 @@
 
 #include <chrono>
 
-#include <catch2/catch.hpp>
+#include <catch2/catch_test_macros.hpp>
 
 #include <BipedalLocomotion/ContactDetectors/SchmittTriggerDetector.h>
 #include <BipedalLocomotion/Math/SchmittTrigger.h>