+++ src/ContinuousDynamicalSystem/tests/IntegratorLinearSystem.cpp
@@ -9,7 +9,7 @@
 #include <memory>
 
 // Catch2
-#include <catch2/catch.hpp>
+#include <catch2/catch_test_macros.hpp>
 
 #include <Eigen/Dense>