+++ src/Conversions/tests/matioCppConversionsTest.cpp
@@ -6,7 +6,7 @@
  */
 
 // Catch2
-#include <catch2/catch.hpp>
+#include <catch2/catch_test_macros.hpp>
 #include <BipedalLocomotion/Conversions/matioCppConversions.h>
 #include <iDynTree/Core/VectorDynSize.h>
 #include <unordered_map>