+++ src/Estimators/tests/FloatingBaseEstimators/InvariantEKFBaseEstimatorTest.cpp
@@ -4,7 +4,7 @@
  * @copyright 2020 Istituto Italiano di Tecnologia (IIT). This software may be modified and
  * distributed under the terms of the BSD-3-Clause license.
  */
-#include <catch2/catch.hpp>
+#include <catch2/catch_test_macros.hpp>
 
 #include <Eigen/Dense>