+++ src/Estimators/tests/RobotDynamicsEstimator/SubModelCreatorTest.cpp
@@ -7,7 +7,7 @@
 
 #include <string>
 // Catch2
-#include <catch2/catch.hpp>
+#include <catch2/catch_test_macros.hpp>
 
 // YARP
 #include <iCubModels/iCubModels.h>