+++ src/GenericContainer/tests/NamedTupleTest.cpp
@@ -11,7 +11,7 @@
 
 // Catch2
 #include <Eigen/src/Core/Matrix.h>
-#include <catch2/catch.hpp>
+#include <catch2/catch_test_macros.hpp>
 
 #include <BipedalLocomotion/GenericContainer/NamedTuple.h>