+++ src/IK/tests/AngularMomentumTaskTest.cpp
@@ -6,7 +6,7 @@
  */
 
 // Catch2
-#include <catch2/catch.hpp>
+#include <catch2/catch_test_macros.hpp>
 
 // BipedalLocomotion
 #include <BipedalLocomotion/IK/AngularMomentumTask.h>