+++ src/ParametersHandler/tests/ParametersHandlerTest.cpp
@@ -12,7 +12,7 @@
 #include <unordered_map>
 
 // Catch2
-#include <catch2/catch.hpp>
+#include <catch2/catch_test_macros.hpp>
 
 #include <BipedalLocomotion/GenericContainer/TemplateHelpers.h>
 #include <BipedalLocomotion/GenericContainer/Vector.h>