+++ src/ParametersHandler/tests/ParametersHandlerTomlTest.cpp
@@ -10,7 +10,7 @@
 #include <memory>
 
 // Catch2
-#include <catch2/catch.hpp>
+#include <catch2/catch_test_macros.hpp>
 
 #include <BipedalLocomotion/GenericContainer/Vector.h>
 #include <BipedalLocomotion/ParametersHandler/IParametersHandler.h>