+++ src/Planners/tests/TimeVaryingDCMPlannerTest.cpp
@@ -8,7 +8,7 @@
 #include <chrono>
 
 // Catch2
-#include <catch2/catch.hpp>
+#include <catch2/catch_test_macros.hpp>
 
 #include <BipedalLocomotion/Contacts/ContactPhaseList.h>
 #include <BipedalLocomotion/Math/Constants.h>