+++ src/RobotInterface/tests/DummyImplementation/DummySensorBridgeTest.cpp
@@ -7,7 +7,7 @@
 
 
 // Catch2
-#include <catch2/catch.hpp>
+#include <catch2/catch_test_macros.hpp>
 #include <BipedalLocomotion/ParametersHandler/IParametersHandler.h>
 #include <BipedalLocomotion/ParametersHandler/StdImplementation.h>