+++ src/SimplifiedModelControllers/tests/CoMZMPControllerTest.cpp
@@ -8,7 +8,7 @@
 #include <array>
 
 // Catch2
-#include <catch2/catch.hpp>
+#include <catch2/catch_test_macros.hpp>
 
 #include <BipedalLocomotion/ParametersHandler/StdImplementation.h>
 #include <BipedalLocomotion/SimplifiedModelControllers/CoMZMPController.h>